+++ src/save.c
@@ -636,7 +636,6 @@ bool load_char_obj( DESCRIPTOR_DATA *d, char *name )
     /* initialize race */
     if (found)
     {
-	int i;
 
 	if (ch->race == 0)
 	    ch->race = race_lookup("human");
@@ -644,12 +643,6 @@ bool load_char_obj( DESCRIPTOR_DATA *d, char *name )
 	ch->size = pc_race_table[ch->race].size;
 	ch->dam_type = 17; /*punch */
 
-	for (i = 0; i < 5; i++)
-	{
-	    if (pc_race_table[ch->race].skills[i] == NULL)
-		break;
-	    group_add(ch,pc_race_table[ch->race].skills[i],FALSE);
-	}
 	ch->affected_by = ch->affected_by|race_table[ch->race].aff;
 	ch->imm_flags	= ch->imm_flags | race_table[ch->race].imm;
 	ch->res_flags	= ch->res_flags | race_table[ch->race].res;
@@ -663,9 +656,6 @@ bool load_char_obj( DESCRIPTOR_DATA *d, char *name )
 
     if (found && ch->version < 2)  /* need to add the new skills */
     {
-	group_add(ch,"rom basics",FALSE);
-	group_add(ch,class_table[ch->class].base_group,FALSE);
-	group_add(ch,class_table[ch->class].default_group,TRUE);
 	ch->pcdata->learned[gsn_recall] = 50;
     }
  
@@ -941,23 +931,6 @@ void fread_char( CHAR_DATA *ch, FILE *fp )
 
 	case 'G':
 	    KEY( "Gold",	ch->gold,		fread_number( fp ) );
-            if ( !str_cmp( word, "Group" )  || !str_cmp(word,"Gr"))
-            {
-                int gn;
-                char *temp;
- 
-                temp = fread_word( fp ) ;
-                gn = group_lookup(temp);
-                /* gn    = group_lookup( fread_word( fp ) ); */
-                if ( gn < 0 )
-                {
-                    fprintf(stderr,"%s",temp);
-                    bugf("Fread_char: unknown group. ");
-                }
-                else
-		    gn_add(ch,gn);
-                fMatch = TRUE;
-            }
 	    break;
 
 	case 'H':